Create Monitor Project Server Queue from data
Project Server 2010 and 2013 has a built in webpage to filter and view the project server queue jobs. But what if you want to keep this data for analysis or automated wornings?
The query below will let you get to the same data you see on the queue's webpage. The data comes from views in the DRAFT database. The major trick is to know that INFO_UID is actually your PROJECT_UID. Using this query you can easily filter on specific states or messages (see the codes) . Use it to spot projects that take long to pubish. You can also spot errors that come from [GRP_QUEUE_ERROR_INFO] field.